% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{add_bar,InputHeatmap-method}
\alias{add_bar,InputHeatmap-method}
\title{add_bar}
\usage{
\S4method{add_bar}{InputHeatmap}(.data, .column, palette = NULL)
}
\value{
A `add_bar` object
}
\description{
add_bar
}
