% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{add_line}
\alias{add_line}
\alias{add_line,InputHeatmap-method}
\title{Adds a line annotation layer to a `InputHeatmap`, that on evaluation creates a `ComplexHeatmap`}
\usage{
add_line(.data, .column, palette = NULL, ...)

\S4method{add_line}{InputHeatmap}(.data, .column, palette = NULL, ...)
}
\arguments{
\item{.data}{A `tbl_df` formatted as | <ELEMENT> | <FEATURE> | <VALUE> | <...> |}

\item{.column}{Vector of quotes}

\item{palette}{A character vector of colors  This is the list of palettes that will be used for horizontal and vertical discrete annotations. The discrete classification of annotations depends on the column type of your input tibble (e.g., character and factor).}

\item{...}{The arguments that will be passed to top_annotation or left_annotation of the ComplexHeatmap container}
}
\value{
A `InputHeatmap` object that gets evaluated to a `ComplexHeatmap`

A `InputHeatmap` object that gets evaluated to a `ComplexHeatmap`
}
\description{
add_line() from a `InputHeatmap` object, adds a line annotation layer.
}
\details{
\lifecycle{maturing}

It uses `ComplexHeatmap` as visualisation tool.
}
\examples{

library(dplyr)

hm = 
  tidyHeatmap::N52 \%>\%
  tidyHeatmap::heatmap(
    .row = symbol_ct,
    .column = UBR,
    .value = `read count normalised log`
)

hm \%>\% add_line()


}
