% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{add_line,InputHeatmap-method}
\alias{add_line,InputHeatmap-method}
\title{add_line}
\usage{
\S4method{add_line}{InputHeatmap}(.data, .column, palette = NULL)
}
\value{
A `add_line` object
}
\description{
add_line
}
