% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{add_tile}
\alias{add_tile}
\alias{add_tile,InputHeatmap-method}
\title{Adds a tile annotation layer to a `InputHeatmap`, that on evaluation creates a `ComplexHeatmap`}
\usage{
add_tile(.data, .column, palette = NULL, size = NULL, ...)

\S4method{add_tile}{InputHeatmap}(.data, .column, palette = NULL, size = NULL, ...)
}
\arguments{
\item{.data}{A `tbl_df` formatted as | <ELEMENT> | <FEATURE> | <VALUE> | <...> |}

\item{.column}{Vector of quotes}

\item{palette}{A character vector of colors, or a function such as colorRamp2 (see examples).}

\item{size}{A grid::unit object, e.g. unit(2, "cm"). This is the height or width of the annotation depending on the orientation.}

\item{...}{The arguments that will be passed to top_annotation or left_annotation of the ComplexHeatmap container}
}
\value{
A `InputHeatmap` object that gets evaluated to a `ComplexHeatmap`

A `InputHeatmap` object that gets evaluated to a `ComplexHeatmap`
}
\description{
add_tile() from a `InputHeatmap` object, adds a tile annotation layer.
}
\details{
\lifecycle{maturing}

It uses `ComplexHeatmap` as visualisation tool.
}
\examples{

library(dplyr)

hm = 
  tidyHeatmap::N52 \%>\%
  tidyHeatmap::heatmap(
    .row = symbol_ct,
    .column = UBR,
    .value = `read count normalised log`
)

hm \%>\% add_tile(CAPRA_TOTAL)


hm \%>\% add_tile(inflection, palette = circlize::colorRamp2(c(0, 3,10), c("white", "green", "red")))

}
