% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{add_tile,InputHeatmap-method}
\alias{add_tile,InputHeatmap-method}
\title{add_tile}
\usage{
\S4method{add_tile}{InputHeatmap}(.data, .column, palette = NULL)
}
\arguments{
\item{.data}{A `tbl_df` formatted as | <ELEMENT> | <FEATURE> | <VALUE> | <...> |}

\item{.column}{Vector of quotes}

\item{palette}{A list of character vectors. This is the list of palettes that will be used for horizontal and vertical discrete annotations. The discrete classification of annotations depends on the column type of your input tibble (e.g., character and factor).}
}
\value{
A `add_tile` object
}
\description{
add_tile
}
