% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{add_tile}
\alias{add_tile}
\title{Creates a  `InputHeatmap` plot from `tbl_df` on evaluation creates a `ComplexHeatmap`}
\usage{
add_tile(.data, .column, palette = NULL)
}
\arguments{
\item{.data}{A `tbl_df` formatted as | <ELEMENT> | <FEATURE> | <VALUE> | <...> |}

\item{.column}{Vector of quotes}

\item{palette}{A list of character vectors. This is the list of palettes that will be used for horizontal and vertical discrete annotations. The discrete classification of annotations depends on the column type of your input tibble (e.g., character and factor).}
}
\value{
A `InputHeatmap` object that gets evaluated to a `ComplexHeatmap`
}
\description{
heatmap() takes a tbl object and easily produces a ComplexHeatmap plot, with integration with tibble and dplyr frameworks.
}
\details{
\lifecycle{maturing}

To be added.
}
\examples{

library(dplyr)
tidyHeatmap::N52 \%>\%
group_by( `Cell type`) \%>\%
tidyHeatmap::heatmap(
 .row = symbol_ct,
 .column = UBR,
 .value = `read count normalised log`,
 annotation = CAPRA_TOTAL
)


}
