% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_if_counts_is_na}
\alias{check_if_counts_is_na}
\title{Check whether there are NA counts}
\usage{
check_if_counts_is_na(.data, .abundance)
}
\arguments{
\item{.data}{A tibble of read counts}

\item{.abundance}{A character name of the read count column}
}
\value{
A tbl
}
\description{
Check whether there are NA counts
}
