% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_exprs}
\alias{data_exprs}
\title{Expression data from RNA-Seq study.}
\format{A data frame with 800 rows and 12 variables:
\describe{
\item{ensembl_gene_id}{...}
\item{external_gene_name}{...}
\item{Eip_vs_Hip.padj}{...}
\item{direction}{...}
\item{is_immune_gene}{...}
\item{count_type}{...}
\item{group}{...}
\item{replicate}{...}
\item{expression}{...}
\item{sample}{...}
\item{sample_type}{...}
\item{condition}{...}
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE104899}
}
\usage{
data_exprs
}
\description{
Some more desciption.
}
\keyword{datasets}
