% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{error_if_duplicated_genes}
\alias{error_if_duplicated_genes}
\title{Check whether there are duplicated genes/transcripts}
\usage{
error_if_duplicated_genes(
  .data,
  .sample = sample,
  .transcript = transcript,
  .abundance = `read count`
)
}
\arguments{
\item{.data}{A tibble of read counts}

\item{.sample}{A character name of the sample column}

\item{.transcript}{A character name of the transcript/gene column}

\item{.abundance}{A character name of the read count column}
}
\value{
A tbl
}
\description{
Check whether there are duplicated genes/transcripts
}
