% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{error_if_log_transformed}
\alias{error_if_log_transformed}
\title{Check whether a numeric vector has been log transformed}
\usage{
error_if_log_transformed(x, .abundance)
}
\arguments{
\item{x}{A numeric vector}

\item{.abundance}{A character name of the transcript/gene abundance column}
}
\value{
NA
}
\description{
Check whether a numeric vector has been log transformed
}
