% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{error_if_wrong_input}
\alias{error_if_wrong_input}
\title{Check whether there are NA counts}
\usage{
error_if_wrong_input(.data, list_input, expected_type)
}
\arguments{
\item{.data}{A tibble of read counts}

\item{list_input}{A list}

\item{expected_type}{A character string}
}
\value{
A tbl
}
\description{
Check whether there are NA counts
}
