% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_abundance_norm_if_exists}
\alias{get_abundance_norm_if_exists}
\title{Get column names either from user or from attributes}
\usage{
get_abundance_norm_if_exists(.data, .abundance)
}
\arguments{
\item{.data}{A tibble}

\item{.abundance}{A character name of the abundance column}
}
\value{
A list of column enquo or error
}
\description{
Get column names either from user or from attributes
}
