% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_elements_features_abundance}
\alias{get_elements_features_abundance}
\title{Get column names either from user or from attributes}
\usage{
get_elements_features_abundance(
  .data,
  .element,
  .feature,
  .abundance,
  of_samples = TRUE
)
}
\arguments{
\item{.data}{A tibble}

\item{.element}{A character name of the sample column}

\item{.feature}{A character name of the transcript/gene column}

\item{.abundance}{A character name of the read count column}

\item{of_samples}{A boolean}
}
\value{
A list of column enquo or error
}
\description{
Get column names either from user or from attributes
}
