% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get_heatmap_order}
\alias{get_heatmap_order}
\alias{get_heatmap_order,InputHeatmap-method}
\title{Retrieve row and column ordering from a heatmap}
\source{
[Mangiola and Papenfuss., 2020](https://joss.theoj.org/papers/10.21105/joss.02472)
}
\usage{
get_heatmap_order(.data)

\S4method{get_heatmap_order}{InputHeatmap}(.data)
}
\arguments{
\item{.data}{A `InputHeatmap` object from tidyHeatmap::heatmap()}
}
\value{
A list with rows and columns in heatmap order

A list containing:
\itemize{
  \item rows: Character vector of row names in heatmap order
  \item columns: Character vector of column names in heatmap order
}
}
\description{
get_heatmap_order() extracts just the row and column names in the order they appear in the heatmap after clustering.
}
\details{
\lifecycle{maturing}

This function is a lighter version of get_ordered_data() that only returns the ordering information without reordering the original data.
}
\examples{

hm <- tidyHeatmap::N52 |>
  tidyHeatmap::heatmap(
    .row = symbol_ct,
    .column = UBR,
    .value = `read count normalised log`
  )

# Get ordering information
order_info <- hm |> get_heatmap_order()
print(order_info$rows)
print(order_info$columns)

}
\references{
Mangiola, S. and Papenfuss, A.T., 2020. "tidyHeatmap: an R package for 
  modular heatmap production based on tidy principles." Journal of Open Source Software.
  doi:10.21105/joss.02472.
}