% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get_ordered_matrix}
\alias{get_ordered_matrix}
\alias{get_ordered_matrix,InputHeatmap-method}
\title{Retrieve the abundance matrix in heatmap order}
\source{
[Mangiola and Papenfuss., 2020](https://joss.theoj.org/papers/10.21105/joss.02472)
}
\usage{
get_ordered_matrix(.data)

\S4method{get_ordered_matrix}{InputHeatmap}(.data)
}
\arguments{
\item{.data}{A `InputHeatmap` object from tidyHeatmap::heatmap()}
}
\value{
Matrix with rows and columns ordered as in the heatmap

A matrix with rows and columns ordered as in the heatmap
}
\description{
get_ordered_matrix() extracts the abundance matrix with rows and columns ordered exactly as they appear in the heatmap.
}
\details{
\lifecycle{maturing}

This function returns the underlying matrix used for the heatmap with rows and columns reordered according to the clustering results.
}
\examples{

hm <- tidyHeatmap::N52 |>
  tidyHeatmap::heatmap(
    .row = symbol_ct,
    .column = UBR,
    .value = `read count normalised log`
  )

# Get ordered matrix
ordered_matrix <- hm |> get_ordered_matrix()
print(dim(ordered_matrix))

}
\references{
Mangiola, S. and Papenfuss, A.T., 2020. "tidyHeatmap: an R package for 
  modular heatmap production based on tidy principles." Journal of Open Source Software.
  doi:10.21105/joss.02472.
}