% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_sample_counts}
\alias{get_sample_counts}
\title{Get column names either from user or from attributes}
\usage{
get_sample_counts(.data, .sample, .abundance)
}
\arguments{
\item{.data}{A tibble}

\item{.sample}{A character name of the sample column}

\item{.abundance}{A character name of the read count column}
}
\value{
A list of column enquo or error
}
\description{
Get column names either from user or from attributes
}
