% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_sample_transcript}
\alias{get_sample_transcript}
\title{Get column names either from user or from attributes}
\usage{
get_sample_transcript(.data, .sample, .transcript)
}
\arguments{
\item{.data}{A tibble}

\item{.sample}{A character name of the sample column}

\item{.transcript}{A character name of the transcript/gene column}
}
\value{
A list of column enquo or error
}
\description{
Get column names either from user or from attributes
}
