% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_x_y_annotation_columns}
\alias{get_x_y_annotation_columns}
\title{get_x_y_annotation_columns}
\usage{
get_x_y_annotation_columns(.data, .horizontal, .vertical, .abundance)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.horizontal}{The name of the column horizontally presented in the heatmap}

\item{.vertical}{The name of the column vertically presented in the heatmap}

\item{.abundance}{The name of the transcript/gene abundance column}
}
\value{
A list
}
\description{
get_x_y_annotation_columns
}
