% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{heatmap}
\alias{heatmap}
\alias{heatmap_}
\alias{heatmap,tbl-method}
\title{Creates a  `InputHeatmap` object from `tbl_df` on evaluation creates a `ComplexHeatmap`}
\usage{
heatmap(
  .data,
  .row,
  .column,
  .value,
  transform = NULL,
  .scale = "row",
  palette_value = c("#440154FF", "#21908CFF", "#fefada"),
  palette_grouping = list(),
  annotation = NULL,
  type = rep("tile", length(quo_names(annotation))),
  palette_discrete = list(),
  palette_continuous = list(),
  .abundance = NULL,
  .horizontal = NULL,
  .vertical = NULL,
  log_transform = NULL,
  palette_abundance = NULL,
  ...
)

heatmap_(
  .data,
  .row,
  .column,
  .value,
  transform = NULL,
  .scale = "row",
  palette_value = c("#440154FF", "#21908CFF", "#fefada"),
  palette_grouping = list(),
  annotation = NULL,
  type = rep("tile", length(quo_names(annotation))),
  palette_discrete = list(),
  palette_continuous = list(),
  .abundance = NULL,
  .horizontal = NULL,
  .vertical = NULL,
  log_transform = NULL,
  palette_abundance = NULL,
  ...
)

\S4method{heatmap}{tbl}(
  .data,
  .row,
  .column,
  .value,
  transform = NULL,
  .scale = "row",
  palette_value = c("#440154FF", "#21908CFF", "#fefada"),
  palette_grouping = list(),
  annotation = NULL,
  type = rep("tile", length(quo_names(annotation))),
  palette_discrete = list(),
  palette_continuous = list(),
  .abundance = NULL,
  .horizontal = NULL,
  .vertical = NULL,
  log_transform = NULL,
  palette_abundance = NULL,
  ...
)
}
\arguments{
\item{.data}{A `tbl_df` formatted as | <ELEMENT> | <FEATURE> | <VALUE> | <...> |}

\item{.row}{The name of the column vertically presented in the heatmap}

\item{.column}{The name of the column horizontally presented in the heatmap}

\item{.value}{The name of the transcript/gene abundance column}

\item{transform}{A function, used to transform .value row-wise (e.g., transform = log1p)}

\item{.scale}{A character string. Possible values are c(\"none\", \"row\", \"column\", \"both\")}

\item{palette_value}{A character vector This is the palette that will be used as gradient for .value. For higher flexibility you can use circlize::colorRamp2\(c\(-2, -1, 0, 1, 2\), viridis::magma\(5\)\)}

\item{palette_grouping}{A list of character vectors. This is the list of palettes that will be used for grouping}

\item{annotation}{DEPRECATED. please use the annotation functions add_* function \(\* one of tile, point, bar, line  \).}

\item{type}{DEPRECATED. please use the annotation functions add_* function \(\* one of tile, point, bar, line  \).}

\item{palette_discrete}{DEPRECATED. please use the annotation functions add_* function \(\* one of tile, point, bar, line  \).}

\item{palette_continuous}{DEPRECATED. please use the annotation functions add_* function \(\* one of tile, point, bar, line  \).}

\item{.abundance}{DEPRECATED. Please use .value instead}

\item{.horizontal}{DEPRECATED. Please use .column instead}

\item{.vertical}{DEPRECATED. Please use .row instead}

\item{log_transform}{DEPRECATED. Please use transform instead}

\item{palette_abundance}{DEPRECATED. Please use palette_value instead}

\item{...}{Further arguments to be passed to ComplexHeatmap::Heatmap}
}
\value{
A `InputHeatmap` objects that gets evaluated to a `ComplexHeatmap` object

A `InputHeatmap` object

A `InputHeatmap` object
}
\description{
heatmap() takes a tbl object and easily produces a ComplexHeatmap plot, with integration with tibble and dplyr frameworks.
}
\details{
\lifecycle{maturing}

This function takes a tbl as an input and creates a `ComplexHeatmap` plot. The information is stored in a `InputHeatmap` object that is updated along the pipe statement, for example adding annotation layers.
}
\examples{

library(dplyr)

tidyHeatmap::N52 \%>\%
group_by( `Cell type`) \%>\%
tidyHeatmap::heatmap(
 .row = symbol_ct,
 .column = UBR,
 .value = `read count normalised log`,
)

}
