% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{heatmap}
\alias{heatmap}
\title{Creates a  `ComplexHeatmap` plot from `tbl`}
\usage{
heatmap(
  .data,
  .horizontal,
  .vertical,
  .abundance,
  annotation = NULL,
  log_transform = FALSE
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.horizontal}{The name of the column horizontally presented in the heatmap}

\item{.vertical}{The name of the column vertically presented in the heatmap}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{annotation}{Vector of quotes}

\item{log_transform}{A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)}
}
\value{
A `ComplexHeatmap` object
}
\description{
plot_heatmap() takes a tbl object and easily produces a ComplexHeatmap plot, with integration with tibble and dplyr frameworks.
}
\details{
To be added.
}
\examples{

library(dplyr)
tidyHeatmap::N52 \%>\%
group_by( `Cell type`) \%>\%
tidyHeatmap::heatmap(
 .horizontal = UBR,
 .vertical = symbol_ct,
 .abundance = `read count normalised log`,
 annotation = CAPRA_TOTAL
)


}
