% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{layer_arrow_up}
\alias{layer_arrow_up}
\alias{layer_arrow_up,InputHeatmap-method}
\alias{layer_arrow_down}
\alias{layer_arrow_down,InputHeatmap-method}
\alias{layer_point}
\alias{layer_point,InputHeatmap-method}
\alias{layer_square}
\alias{layer_square,InputHeatmap-method}
\alias{layer_diamond}
\alias{layer_diamond,InputHeatmap-method}
\title{Adds a layers of symbols above the heatmap tiles to a `InputHeatmap`, that on evaluation creates a `ComplexHeatmap`}
\usage{
layer_arrow_up(.data, ...)

\S4method{layer_arrow_up}{InputHeatmap}(.data, ...)

layer_arrow_down(.data, ...)

\S4method{layer_arrow_down}{InputHeatmap}(.data, ...)

layer_point(.data, ...)

\S4method{layer_point}{InputHeatmap}(.data, ...)

layer_square(.data, ...)

\S4method{layer_square}{InputHeatmap}(.data, ...)

layer_diamond(.data, ...)

\S4method{layer_diamond}{InputHeatmap}(.data, ...)
}
\arguments{
\item{.data}{A `InputHeatmap`}

\item{...}{Expressions that return a logical value, and are defined in terms of the variables in .data. If multiple expressions are included, they are combined with the & operator. Only rows for which all conditions evaluate to TRUE are kept.}
}
\value{
A `InputHeatmap` object that gets evaluated to a `ComplexHeatmap`

A `InputHeatmap` object that gets evaluated to a `ComplexHeatmap`

A `InputHeatmap` object that gets evaluated to a `ComplexHeatmap`

A `InputHeatmap` object that gets evaluated to a `ComplexHeatmap`

A `InputHeatmap` object that gets evaluated to a `ComplexHeatmap`

A `InputHeatmap` object that gets evaluated to a `ComplexHeatmap`

A `InputHeatmap` object that gets evaluated to a `ComplexHeatmap`

A `InputHeatmap` object that gets evaluated to a `ComplexHeatmap`

A `InputHeatmap` object that gets evaluated to a `ComplexHeatmap`

A `InputHeatmap` object that gets evaluated to a `ComplexHeatmap`
}
\description{
layer_arrow_up() from a `InputHeatmap` object, adds a bar annotation layer.

layer_arrow_down() from a `InputHeatmap` object, adds a bar annotation layer.

layer_point() from a `InputHeatmap` object, adds a bar annotation layer.

layer_square() from a `InputHeatmap` object, adds a bar annotation layer.

layer_diamond() from a `InputHeatmap` object, adds a bar annotation layer.
}
\details{
\lifecycle{maturing}

It uses `ComplexHeatmap` as visualisation tool.

\lifecycle{maturing}

It uses `ComplexHeatmap` as visualisation tool.

\lifecycle{maturing}

It uses `ComplexHeatmap` as visualisation tool.

\lifecycle{maturing}

It uses `ComplexHeatmap` as visualisation tool.

\lifecycle{maturing}

It uses `ComplexHeatmap` as visualisation tool.
}
\examples{

library(dplyr)

hm = 
  tidyHeatmap::N52 \%>\%
  tidyHeatmap::heatmap(
    .row = symbol_ct,
    .column = UBR,
    .value = `read count normalised log`
)

hm \%>\% layer_arrow_up()



library(dplyr)

hm = 
  tidyHeatmap::N52 \%>\%
  tidyHeatmap::heatmap(
    .row = symbol_ct,
    .column = UBR,
    .value = `read count normalised log`
)

hm \%>\% layer_arrow_down()



library(dplyr)

hm = 
  tidyHeatmap::N52 \%>\%
  tidyHeatmap::heatmap(
    .row = symbol_ct,
    .column = UBR,
    .value = `read count normalised log`
)

hm \%>\% layer_point()



library(dplyr)

hm = 
  tidyHeatmap::N52 \%>\%
  tidyHeatmap::heatmap(
    .row = symbol_ct,
    .column = UBR,
    .value = `read count normalised log`
)

hm \%>\% layer_square()



library(dplyr)

hm = 
  tidyHeatmap::N52 \%>\%
  tidyHeatmap::heatmap(
    .row = symbol_ct,
    .column = UBR,
    .value = `read count normalised log`
)

hm \%>\% layer_diamond()


}
