% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{layer_arrow_down,InputHeatmap-method}
\alias{layer_arrow_down,InputHeatmap-method}
\title{layer_arrow_down}
\usage{
\S4method{layer_arrow_down}{InputHeatmap}(.data, ...)
}
\value{
A `layer_arrow_down` object
}
\description{
layer_arrow_down
}
