% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{layer_arrow_up,InputHeatmap-method}
\alias{layer_arrow_up,InputHeatmap-method}
\title{layer_arrow_up}
\usage{
\S4method{layer_arrow_up}{InputHeatmap}(.data, ...)
}
\value{
A `layer_arrow_up` object
}
\description{
layer_arrow_up
}
