% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{layer_point,InputHeatmap-method}
\alias{layer_point,InputHeatmap-method}
\title{layer_point}
\usage{
\S4method{layer_point}{InputHeatmap}(.data, ...)
}
\value{
A `layer_point` object
}
\description{
layer_point
}
