% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{N52}
\alias{N52}
\title{Example data set N52}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 520 rows and 15 columns.
}
\source{
[Mangiola and Papenfuss., 2020](https://joss.theoj.org/papers/10.21105/joss.02472)
}
\usage{
N52
}
\description{
Example data set N52
}
\references{
Mangiola, S. and Papenfuss, A.T., 2020. "tidyHeatmap: an R package for 
  modular heatmap production based on tidy principles." Journal of Open Source Software.
  doi:10.21105/joss.02472.
}
\keyword{datasets}
