% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{+.InputHeatmap}
\alias{+.InputHeatmap}
\title{Add two tidyHeatmap objects together}
\source{
[Mangiola and Papenfuss., 2020](https://joss.theoj.org/papers/10.21105/joss.02472)
}
\usage{
\method{+}{InputHeatmap}(e1, e2)
}
\arguments{
\item{e1}{First tidyHeatmap object}

\item{e2}{Second tidyHeatmap object}
}
\value{
A combined ComplexHeatmap object
}
\description{
Add two tidyHeatmap objects together
}
\references{
Mangiola, S. and Papenfuss, A.T., 2020. "tidyHeatmap: an R package for 
  modular heatmap production based on tidy principles." Journal of Open Source Software.
  doi:10.21105/joss.02472.
}
