% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{prepend}
\alias{prepend}
\title{From rlang deprecated}
\usage{
prepend(x, values, before = 1)
}
\arguments{
\item{x}{An array}

\item{values}{An array}

\item{before}{A boolean}
}
\value{
An array
}
\description{
From rlang deprecated
}
