% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{quo_names}
\alias{quo_names}
\title{Convert array of quosure (e.g. c(col_a, col_b)) into character vector}
\usage{
quo_names(v)
}
\arguments{
\item{v}{A array of quosures (e.g. c(col_a, col_b))}
}
\value{
A character vector
}
\description{
Convert array of quosure (e.g. c(col_a, col_b)) into character vector
}
