% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{unit}
\alias{colorRamp2}
\title{Re-export selected functions for user convenience}
\description{
These re-exports allow using `unit` and `colorRamp2` without
namespace qualification after attaching tidyHeatmap.
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{circlize}{\code{\link[circlize]{colorRamp2}}}

  \item{grid}{\code{\link[grid]{unit}}}
}}

