% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{save_pdf,Heatmap-method}
\alias{save_pdf,Heatmap-method}
\title{save_pdf}
\usage{
\S4method{save_pdf}{Heatmap}(
  .heatmap,
  filename,
  width = NULL,
  height = NULL,
  units = c("in", "cm", "mm")
)
}
\description{
save_pdf
}
