% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{save_pdf,InputHeatmap-method}
\alias{save_pdf,InputHeatmap-method}
\title{save_pdf}
\usage{
\S4method{save_pdf}{InputHeatmap}(
  .heatmap,
  filename,
  width = NULL,
  height = NULL,
  units = c("in", "cm", "mm")
)
}
\arguments{
\item{.heatmap}{A `Heatmap`}

\item{filename}{A character string. The name of the output file/path}

\item{width}{A `double`. Plot width}

\item{height}{A `double`. Plot height}

\item{units}{A character string. units ("in", "cm", or "mm")}
}
\description{
save_pdf
}
