% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{save_pdf}
\alias{save_pdf}
\title{Save plot on PDF file}
\usage{
save_pdf(
  .heatmap,
  filename,
  width = NULL,
  height = NULL,
  units = c("in", "cm", "mm")
)
}
\arguments{
\item{.heatmap}{A `Heatmap`}

\item{filename}{A character string. The name of the output file/path}

\item{width}{A `double`. Plot width}

\item{height}{A `double`. Plot height}

\item{units}{A character string. units ("in", "cm", or "mm")}
}
\value{
NA
}
\description{
save_pdf() takes as input a Heatmap from ComplexHeatmap and save it to PDF file
}
\details{
\lifecycle{maturing}

It simply save an `Heatmap` to a PDF file use pdf() function in the back end
}
\examples{


	tidyHeatmap::heatmap(
  dplyr::group_by(tidyHeatmap::pasilla,		location, type),
  .column = sample,
  .row = symbol,
  .value = `count normalised adjusted`,
 ) |>
 save_pdf(tempfile())


}
