% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{scale_design}
\alias{scale_design}
\title{Scale design matrix}
\usage{
scale_design(df, .formula)
}
\arguments{
\item{df}{A tibble}

\item{.formula}{a formula}
}
\value{
A tibble
}
\description{
Scale design matrix
}
