% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{scale_robust}
\alias{scale_robust}
\title{Scale counts in a robust way against sd == 0}
\usage{
scale_robust(y)
}
\arguments{
\item{y}{A numerical array}
}
\value{
A scaled and centred numerical array
}
\description{
Scale counts in a robust way against sd == 0
}
