% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{select_closest_pairs}
\alias{select_closest_pairs}
\title{Sub function of remove_redundancy_elements_though_reduced_dimensions}
\usage{
select_closest_pairs(df)
}
\arguments{
\item{df}{A tibble}
}
\value{
A tibble with pairs to drop
}
\description{
Sub function of remove_redundancy_elements_though_reduced_dimensions
}
