% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyheatmap.R
\name{tidy_heatmap}
\alias{tidy_heatmap}
\title{Plot pretty heatmaps from tidy data}
\usage{
tidy_heatmap(df, rows, columns, values, filename = NA, ann_row = NULL,
  ann_col = NULL, gaps_row = NULL, gaps_col = NULL,
  ann_colors = NA, breaks = NA, scale = "none", fontsize = 7,
  cellwidth = NA, cellheight = NA, color_scale_min = NA,
  color_scale_max = NA, color_scale_n = 100,
  color = (grDevices::colorRampPalette(rev(RColorBrewer::brewer.pal(n =
  7, name = "RdYlBu"))))(color_scale_n), return_data = FALSE,
  cluster_rows = FALSE, cluster_cols = FALSE, border_color = NA, ...)
}
\arguments{
\item{df}{Input dataframe in long format.}

\item{rows}{Column of the dataframe to use for heatmap rows.}

\item{columns}{Column of the dataframe to use for heatmap columns.}

\item{values}{Column of the dataframe containing the values to plot.}

\item{filename}{todo}

\item{ann_row}{todo}

\item{ann_col}{todo}

\item{gaps_row}{todo}

\item{gaps_col}{todo}

\item{ann_colors}{List of colors used to color code row and column annotations.
See \code{pheatmap::pheatmap()} for details.}

\item{breaks}{Sequence used for mapping values to colors. Is one element longer than \code{color_scale_n}.
See \code{pheatmap::pheatmap()} for details.}

\item{scale}{Character indicating if the values should be centered and scaled by "row", "column" or "none".
Default is "row".}

\item{fontsize}{Base fontsize for the plot.}

\item{cellwidth}{todo}

\item{cellheight}{todo}

\item{color_scale_min}{todo}

\item{color_scale_max}{todo}

\item{color_scale_n}{Number of color on color scale.}

\item{color}{Vector of colors used in heatmap.}

\item{return_data}{If \code{TRUE} processed data is returned instead of plot.}

\item{cluster_rows}{todo}

\item{cluster_cols}{todo}

\item{border_color}{todo}

\item{...}{Other parameters passed to \code{pheatmap::pheatmap()}.}
}
\description{
This function is a wrapper around \code{pheatmap::pheatmap()} that allows long tidy
data as input.
}
