% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyHeatmap-package.R
\docType{package}
\name{tidyHeatmap-package}
\alias{tidyHeatmap}
\alias{tidyHeatmap-package}
\title{tidyHeatmap: A Tidy Implementation of Heatmap}
\description{
This is a tidy implementation for heatmap. At the moment it is based on the 
(great) package 'ComplexHeatmap'. The goal of this package is to interface a 
tidy data frame with this powerful tool. Some of the advantages are: Row and/or 
columns colour annotations are easy to integrate just specifying one parameter 
(column names). Custom grouping of rows is easy to specify providing a grouped 
tbl. For example: df \%>\% group_by(...). Labels size adjusted by row and column 
total number. Default use of Brewer and Viridis palettes.
}
\details{
The tidyHeatmap package provides a tidy interface to create information-rich 
heatmaps using the powerful ComplexHeatmap engine. The package integrates 
seamlessly with the tidyverse ecosystem, allowing you to use familiar dplyr 
and tidyr verbs.

Main functions:
\itemize{
  \item \code{\link{heatmap}}: Create a heatmap from a tidy data frame
  \item \code{\link{annotation_group}}: Add group annotation strips
  \item \code{\link{annotation_tile}}: Add tile annotations
  \item \code{\link{annotation_point}}: Add point annotations
  \item \code{\link{annotation_bar}}: Add bar annotations
  \item \code{\link{annotation_numeric}}: Add numeric annotations
  \item \code{\link{annotation_line}}: Add line annotations
  \item \code{\link{layer_text}}: Add text layers
  \item \code{\link{layer_point}}: Add point layers
  \item \code{\link{layer_square}}: Add square symbol layers
  \item \code{\link{layer_diamond}}: Add diamond symbol layers
  \item \code{\link{layer_arrow_up}}: Add arrow up symbol layers
  \item \code{\link{layer_arrow_down}}: Add arrow down symbol layers
  \item \code{\link{layer_asterisk}}: Add asterisk symbol layers
  \item \code{\link{layer_star}}: Add star symbol layers
  \item \code{\link{get_heatmap_data}}: Extract matrix and dendrograms from heatmap
  \item \code{\link{as_ComplexHeatmap}}: Convert to ComplexHeatmap object
  \item \code{\link{save_pdf}}: Save heatmap to PDF
  \item \code{\link{scale_robust}}: Robust scaling function
  \item \code{\link{split_columns}}: Split columns in heatmap
}

The package uses a pipe-friendly syntax that allows you to build complex 
heatmaps by adding layers of annotations and customizations step by step.
}
\examples{
# Basic heatmap
tidyHeatmap::N52 |>
  heatmap(
    .row = symbol_ct,
    .column = UBR,
    .value = `read count normalised log`
  )

# Heatmap with annotations
tidyHeatmap::pasilla |>
  heatmap(
    .column = sample,
    .row = symbol,
    .value = `count normalised adjusted`,
    scale = "row"
  ) |>
  annotation_tile(condition) |>
  annotation_point(activation)

}
\references{
Mangiola, S. and Papenfuss, A.T., 2020. "tidyHeatmap: an R package for 
  modular heatmap production based on tidy principles." Journal of Open Source Software, 
  5(52), 2472. \doi{10.21105/joss.02472}

Gu, Z., Eils, R., and Schlesner, M. 2016. "Complex heatmaps reveal patterns 
  and correlations in multidimensional genomic data." Bioinformatics, 32(18), 2847-2849.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stemangiola/tidyHeatmap}
  \item Report bugs at \url{https://github.com/stemangiola/tidyHeatmap/issues}
}

}
\author{
\strong{Maintainer}: Stefano Mangiola \email{mangiolastefano@gmail.com}

Other contributors:
\itemize{
  \item Anthony Papenfuss \email{papenfuss@wehi.edu.au} [contributor]
}

}
\keyword{internal}
