% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{add_bar,InputHeatmap-method}
\alias{add_bar,InputHeatmap-method}
\title{add_bar}
\usage{
\S4method{add_bar}{InputHeatmap}(.data, .column, palette = NULL)
}
\arguments{
\item{.data}{A `tbl_df` formatted as | <ELEMENT> | <FEATURE> | <VALUE> | <...> |}

\item{.column}{Vector of quotes}

\item{palette}{A character vector of colors  This is the list of palettes that will be used for horizontal and vertical discrete annotations. The discrete classification of annotations depends on the column type of your input tibble (e.g., character and factor).}
}
\value{
A `add_bar` object
}
\description{
add_bar
}
