% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{annotation_group}
\alias{annotation_group}
\title{Add group annotation strips to a tidyHeatmap}
\usage{
annotation_group(
  .data,
  ...,
  palette_grouping = list(),
  group_label_fontsize = 8,
  group_name_fontsize = 8,
  show_group_name = TRUE,
  group_strip_height = grid::unit(9, "pt"),
  group_strip_width = grid::unit(9, "pt")
)
}
\arguments{
\item{.data}{A tidyHeatmap object}

\item{...}{Grouping columns (unquoted, like dplyr::group_by)}

\item{palette_grouping}{List of color vectors for each grouping}

\item{group_label_fontsize}{Font size for group labels}

\item{group_name_fontsize}{Font size for group annotation name}

\item{show_group_name}{Logical, show the group annotation name}

\item{group_strip_height}{Height of group strip (unit)}

\item{group_strip_width}{Width of column group strip (unit)}
}
\value{
A tidyHeatmap object with group annotation
}
\description{
Add group annotation strips to a tidyHeatmap
}
