% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{annotation_group}
\alias{annotation_group}
\title{Add group annotation strips to a tidyHeatmap}
\usage{
annotation_group(.data, ...)
}
\arguments{
\item{.data}{A tidyHeatmap object}

\item{...}{Grouping columns (unquoted, like dplyr::group_by)}

\item{palette_grouping}{List of color vectors for each grouping}

\item{group_label_fontsize}{Font size for group labels}

\item{show_group_name}{Logical, show the group annotation name}

\item{group_strip_height}{Height of group strip (unit)}
}
\value{
A tidyHeatmap object with group annotation
}
\description{
Add group annotation strips to a tidyHeatmap
}
