% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{as_ComplexHeatmap}
\alias{as_ComplexHeatmap}
\alias{as_ComplexHeatmap,InputHeatmap-method}
\title{Creates a  `ComplexHeatmap` object for less standard plot manipulation (e.g. changing legend position)}
\source{
[Mangiola and Papenfuss., 2020](https://joss.theoj.org/papers/10.21105/joss.02472)

[Mangiola and Papenfuss., 2020](https://joss.theoj.org/papers/10.21105/joss.02472)
}
\usage{
as_ComplexHeatmap(tidyHeatmap)

\S4method{as_ComplexHeatmap}{InputHeatmap}(tidyHeatmap)
}
\arguments{
\item{tidyHeatmap}{A `InputHeatmap` object from tidyHeatmap::heatmap() call}
}
\value{
A `ComplexHeatmap`
}
\description{
as_ComplexHeatmap() takes a `InputHeatmap` object and produces a `Heatmap` object
}
\details{
\lifecycle{maturing}
}
\examples{


tidyHeatmap::N52 |>
tidyHeatmap::heatmap(
 .row = symbol_ct,
 .column = UBR,
 .value = `read count normalised log`,
) |>
as_ComplexHeatmap()

}
\references{
Mangiola, S. and Papenfuss, A.T., 2020. "tidyHeatmap: an R package for 
  modular heatmap production based on tidy principles." Journal of Open Source Software.
  doi:10.21105/joss.02472.

Mangiola, S. and Papenfuss, A.T., 2020. "tidyHeatmap: an R package for 
  modular heatmap production based on tidy principles." Journal of Open Source Software.
  doi:10.21105/joss.02472.
}
