% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{as_ComplexHeatmap}
\alias{as_ComplexHeatmap}
<<<<<<< HEAD
\alias{as_ComplexHeatmap,InputHeatmap-method}
\title{Creates a  `ComplexHeatmap` object for less standard plot manipulation (e.g. changing legend position)}
\usage{
as_ComplexHeatmap(tidyHeatmap)

\S4method{as_ComplexHeatmap}{InputHeatmap}(tidyHeatmap)
=======
\title{Creates a  `ComplexHeatmap` object for less standard plot manipulation (e.g. changing legend position)}
\usage{
as_ComplexHeatmap(tidyHeatmap)
>>>>>>> add as_ComplexHeatmap for more unusual operations with draw
}
\arguments{
\item{tidyHeatmap}{A `InputHeatmap` object from tidyHeatmap::heatmap() call}
}
\value{
A `ComplexHeatmap`
}
\description{
as_ComplexHeatmap() takes a `InputHeatmap` object and produces a `Heatmap` object
}
\details{
\lifecycle{maturing}
}
\examples{


tidyHeatmap::N52 |>
tidyHeatmap::heatmap(
 .row = symbol_ct,
 .column = UBR,
 .value = `read count normalised log`,
) |>
as_ComplexHeatmap()

}
