% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_if_duplicated_genes}
\alias{check_if_duplicated_genes}
\title{Check whether there are duplicated genes/transcripts}
\usage{
check_if_duplicated_genes(.data, .sample, .transcript, .abundance)
}
\arguments{
\item{.data}{A tibble of read counts}

\item{.sample}{A character name of the sample column}

\item{.transcript}{A character name of the transcript/gene column}

\item{.abundance}{A character name of the read count column}
}
\value{
A tbl
}
\description{
Check whether there are duplicated genes/transcripts
}
