% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{heatmap}
\alias{heatmap}
\title{Creates a  `ComplexHeatmap` plot from `tbl`}
\usage{
heatmap(
  .data,
  .horizontal,
  .vertical,
  .abundance,
  annotation = NULL,
  log_transform = F
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.abundance}{The name of the transcript/gene abundance column}

\item{.sample}{The name of the sample column}

\item{.transcript}{The name of the transcript/gene column}
}
\value{
A `heatmap` object
}
\description{
plot_heatmap() takes a tbl object and easily produces a ComplexHeatmap plot, with integration with tibble and dplyr frameworks.
}
\details{
This function created a heatmap object and is useful if you want
to avoid to specify .sample, .transcript and .abundance arguments all the times.
The heatmap object have an attribute called parameters where these three
arguments are stored as metadata. They can be extracted as attr(<object>, "parameters").
}
\examples{

library(tidyverse)
tidyHeatmap::N52 \%>\%
group_by( `Cell type`) \%>\%
tidyHeatmap::plot_heatmap(
 .horizontal = UBR,
 .vertical = symbol_ct,
 .abundance = `read count normalised log`,
 annotation = CAPRA_TOTAL
)


}
