% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{layer_arrow_up,InputHeatmap-method}
\alias{layer_arrow_up,InputHeatmap-method}
\title{layer_arrow_up}
\usage{
\S4method{layer_arrow_up}{InputHeatmap}(.data, ...)
}
\arguments{
\item{.data}{A `InputHeatmap`}

\item{...}{Expressions that return a logical value, and are defined in terms of the variables in .data. If multiple expressions are included, they are combined with the & operator. Only rows for which all conditions evaluate to TRUE are kept.}
}
\value{
A `layer_arrow_up` object
}
\description{
layer_arrow_up
}
