% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{layer_diamond,InputHeatmap-method}
\alias{layer_diamond,InputHeatmap-method}
\title{layer_diamond}
\usage{
\S4method{layer_diamond}{InputHeatmap}(.data, ...)
}
\value{
A `layer_diamond` object
}
\description{
layer_diamond
}
