% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{layer_square,InputHeatmap-method}
\alias{layer_square,InputHeatmap-method}
\title{layer_square}
\usage{
\S4method{layer_square}{InputHeatmap}(.data, ...)
}
\value{
A `layer_square` object
}
\description{
layer_square
}
