% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{layer_text}
\alias{layer_text}
\title{Adds a layers of texts above the heatmap tiles to a `InputHeatmap`, that on evaluation creates a `ComplexHeatmap`}
\usage{
layer_text(.data, ..., .value, .size = NULL)
}
\arguments{
\item{.data}{A `InputHeatmap`}

\item{...}{Expressions that return a logical value, and are defined in terms of the variables in .data. If multiple expressions are included, they are combined with the & operator. Only rows for which all conditions evaluate to TRUE are kept.}

\item{.value}{A column name or character string.}

\item{.size}{A column name or a double.}
}
\value{
A `InputHeatmap` object that gets evaluated to a `ComplexHeatmap`
}
\description{
layer_text() from a `InputHeatmap` object, adds a text annotation layer.
}
\details{
\lifecycle{maturing}

It uses `ComplexHeatmap` as visualisation tool.
}
\examples{

library(dplyr)

hm = 
  tidyHeatmap::N52 |>
  mutate(my_text = "t")
  tidyHeatmap::heatmap(
    .row = text_ct,
    .column = UBR,
    .value = `read count normalised log`
) 

hm |> layer_text(.value = "a")
hm |> layer_text(.value = my_text)

}
