% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{N52}
\alias{N52}
\title{Example data set N52}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6370 rows and 15 columns.}
\usage{
N52
}
\description{
Example data set N52
}
\keyword{datasets}
