% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pasilla}
\alias{pasilla}
\title{Example data set Pasilla}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 504 rows and 7 columns.}
\usage{
pasilla
}
\description{
Example data set Pasilla
}
\keyword{datasets}
