% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(.data, .x, .y, .abundance)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.x}{The name of the column horizontally presented in the heatmap}

\item{.y}{The name of the column vertically presented in the heatmap}

\item{.abundance}{The name of the transcript/gene abundance column}
}
\value{
A `ComplexHeatmap` object
}
\description{
plot_heatmap() creates a `tt` object from a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |
}
\details{
To be added.
}
\examples{


print(1)


}
